fix quill editor for custom fields

+++ app/packs/src/decidim/decidim_awesome/awesome_application.js
@@ -4,12 +4,14 @@ import "src/decidim/decidim_awesome/forms/autosave"
 import {destroyQuillEditor,createQuillEditor,createMarkdownEditor} from "src/decidim/decidim_awesome/editors/editor"
 
 $(() => {
-  $(".editor-container").each((_idx, container) => {
-  	destroyQuillEditor(container);
-    if(window.DecidimAwesome.use_markdown_editor) {
-      createMarkdownEditor(container);
-    } else {
-      createQuillEditor(container);
-    }
-  });
+  if(window.DecidimAwesome.allow_images_in_full_editor || window.DecidimAwesome.allow_images_in_small_editor || window.DecidimAwesome.use_markdown_editor) {
+    $(".editor-container").each((_idx, container) => {
+    	destroyQuillEditor(container);
+      if(window.DecidimAwesome.use_markdown_editor) {
+        createMarkdownEditor(container);
+      } else {
+        createQuillEditor(container);
+      }
+    });
+  }
 });
